pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solana-moon-scanner"
version = "1.0.0"
description = "Production-quality Solana DEX token monitoring and scoring system"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Zayd", email = "zayd@example.com"}
]
keywords = ["solana", "defi", "token-scanner", "moon-score", "dex-monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "solders>=0.18.0",
    "solana>=0.32.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-telegram-bot>=20.7",
    "requests>=2.31.0",
    "asyncio>=3.4.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "aiodns>=3.1.0",
    "prometheus-client>=0.19.0",
    "tenacity>=8.2.3",
    "rich>=13.7.0",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.1",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=6.1.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
moon-scanner = "src.cli:main"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
